Ensure the entire development harness works
I used the Rakefile that I use for many projects, but most projects
don't use Appraisals to test compatibility with different versions of a
gem. I discovered that the Rakefile didn't work with the Appraisals set
up so this change fixes the Rakefile to work consistently with or
without Appraisals.