Use built-in types for both ActiveRecord types
This was a mistake that didn't show as a problem when we were only
testing on SQLite. I believe that, because SQLite has very few types,
the original definition here did not cause problems because everything
was getting loosely typed. When using the gem in PostgreSQL, however, it
was attempting to use the type specified here when using JRuby. Unless
you had defined a custom type in your database, this wouldn't work in
that case.

I find it odd, though, that the problem does _not_ appear in MRI and
only appears in JRuby.