Test on MySQL as well as PostgreSQL and SQLite
MySQL is the last of the popular databases in the Ruby ecosystem that we
want to test on. This change makes it so we test on all three of the
main databases that people use.

This change applies a backport to Rails 5.0 because I want to keep the
compatibility as it was in the original gem but modern MySQL doesn't
work with Rails 5.0. I extracted the patch from the 5-1-stable branch
and minimized it because the commit also touches a file that does not
yet exist.